+++ readline/readline/search.c
@@ -74,7 +74,7 @@ int _rl_search_case_fold = 0;
 /* end_clink_change */
 
 static char *history_search_string;
-static int history_string_size;
+static size_t history_string_size;
 
 static void make_history_line_current (HIST_ENTRY *);
 static int noninc_search_from_pos (char *, int, int, int, int *);
